fix: snippet items trigger errors on accept
For snippet items, the value of the environment variable is used as the
item's `insertText`.

Errors are triggered if the snippet item's `insertText` contains lsp
snippets special constructs like `$`. This is due to blink.cmp trying to
parse them as lsp snippets.

With this commit, snippet item's `insertTextFormat` is set to
`PlainText`. Thus telling blink.cmp to treat the `insertText` as is and
not parsing it as lsp snippets syntax.

Closes: #8